refactor: parsers를 storage 하위로 이동 (storage에서만 사용됨)
